chore(webapp): refactor CreateApp and move api call to common api calls file
